requirements.txt
click==7.1.2
Flask==1.1.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
onetimepass==1.0.1
python-dotenv==0.20.0
six==1.16.0
Werkzeug==1.0.1
